% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_funcs.R
\name{bladderMALDIRompp2010}
\alias{bladderMALDIRompp2010}
\title{Load the example MALDI-MSI data.}
\usage{
bladderMALDIRompp2010(verbose = TRUE)
}
\arguments{
\item{verbose}{Logical (default = TRUE). Show additional output text.}
}
\value{
desiData MS intensity matrix. Rows represent pixels, columns represent
matched peaks.
}
\description{
Loads a single mouse urinary bladder MALDI mass spectrometry imaging dataset
acquired in positive ionization mode using Thermo qExactive Orbitrap. The
dataset is available at
"https://raw.github.com/paoloinglese/SPUTNIKexamples/master/data/bladder_maldi_prepr_MALDIquant.RData"
The dataset is loaded in the R environment under the variable name \code{maldiData}.
}
\references{
Rompp, A., Guenther, S., Schober, Y., Schulz, O., Takats, Z., Kummer, W.,
& Spengler, B. (2010). Histology by mass spectrometry: label-free tissue
characterization obtained from high-accuracy bioanalytical imaging. Angewandte
chemie international edition, 49(22), 3834-3838.
}
