% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms.image_class.R
\docType{class}
\name{ms.image-class}
\alias{ms.image-class}
\title{\link{ms.image-class} definition.}
\description{
\link{ms.image-class} definition.
}
\section{Slots}{

\describe{
\item{\code{values}}{numeric 2-D matrix representing the pixel intensity values.}

\item{\code{name}}{string. Image name used for plotting.}

\item{\code{scaled}}{logical. Whether the pixels intensities have been scaled in [0, 1]
or not.}
}}

\author{
Paolo Inglese \email{p.inglese14@imperial.ac.uk}
}
