% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msi.dataset_class.R
\docType{class}
\name{msi.dataset-class}
\alias{msi.dataset-class}
\title{\link{msi.dataset-class} S4 class definition containing the information about
the mass spectrometry imaging dataset.}
\description{
\link{msi.dataset-class} S4 class definition containing the information about
the mass spectrometry imaging dataset.
}
\section{Slots}{

\describe{
\item{\code{matrix}}{the peaks intensity matrix. Rows represent pixels, and columns
represent peaks.}

\item{\code{mz}}{vector of matched m/z values.}

\item{\code{nrow}}{geometrical shape (number of rows) of image.}

\item{\code{ncol}}{geometrical shape (number of columns) of image.}

\item{\code{norm}}{normalization method.}

\item{\code{normoffset}}{numeric offset used for the normalization.}

\item{\code{vartr}}{variance stabilizing transformation.}

\item{\code{vartroffset}}{numeric offset used for the variance stabilizing transformation.}

\item{\code{numdetected}}{msImage of number of detected peaks.}

\item{\code{totalioncount}}{msImage of total-ion-count per pixel.}
}}

\author{
Paolo Inglese \email{p.inglese14@imperial.ac.uk}
}
