% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msi.dataset_methods.R
\name{numDetectedMSI,msi.dataset-method}
\alias{numDetectedMSI,msi.dataset-method}
\alias{numDetectedMSI}
\title{Generates an msImage representing the number of detected peaks per pixel. This
image can be used to qualitatively evaluate the spatial heterogeneity of the
sample.}
\usage{
\S4method{numDetectedMSI}{msi.dataset}(object)
}
\arguments{
\item{object}{\link{msi.dataset-class} object.}
}
\value{
\link{ms.image-class} object representing the detected ions per pixel.
}
\description{
Generates an msImage representing the number of detected peaks per pixel. This
image can be used to qualitatively evaluate the spatial heterogeneity of the
sample.
}
