% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_funcs.R
\name{ovarianDESIDoria2016}
\alias{ovarianDESIDoria2016}
\title{Load the example DESI-MSI data.}
\usage{
ovarianDESIDoria2016(verbose = TRUE)
}
\arguments{
\item{verbose}{Logical (default = TRUE). Show additional output text.}
}
\value{
maldiData MS intensity matrix. Rows represent pixels, columns represent
matched peaks.
}
\description{
Loads a single human ovarian cancer DESI mass spectrometry imaging dataset
acquired in negative ionization mode using Waters XEVO-GS2 qToF. The
dataset is available at
"https://raw.github.com/paoloinglese/SPUTNIKexamples/master/data/ovarian_xevo_prepr_MALDIquant.RData"
The dataset is loaded in the R environment under the variable name \code{maldiData}.
}
\references{
Doria, M. L., McKenzie, J. S., Mroz, A., Phelps, D. L., Speller,
A., Rosini, F., ... & Ghaem-Maghami, S. (2016). Epithelial ovarian carcinoma
diagnosis by desorption electrospray ionization mass spectrometry imaging.
Scientific reports, 6, 39219.
}
