% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms.image_methods.R
\name{plot,ms.image,missing-method}
\alias{plot,ms.image,missing-method}
\alias{plot}
\title{Visualize an MS image.
\code{plot} extends the generic function to \link{ms.image-class} objects.}
\usage{
\S4method{plot}{ms.image,missing}(x, palette = "inferno")
}
\arguments{
\item{x}{\link{ms.image-class} object. See \link{msImage}.}

\item{palette}{string. Color palette. @seealso \code{\link[viridis]{viridis}}.}
}
\value{
a ggplot2 plot.
}
\description{
Visualize an MS image.
\code{plot} extends the generic function to \link{ms.image-class} objects.
}
\examples{
## Load package

\donttest{
  library("SPUTNIK")

  ## Create ms.image-class object
  msIm <- msImage(values = matrix(rnorm(200), 40, 50), name = "test", scale = TRUE)
  
  ## Plot the image
  plot(msIm)
}
}
