% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_refimg.R
\name{refImageBinaryKmeans}
\alias{refImageBinaryKmeans}
\alias{refImageOtsu}
\title{Calculate the binary reference image using k-means clustering. K-Means is run
on the first `npcs` principal components to speed up the calculations.}
\usage{
refImageBinaryKmeans(
  dataset,
  npcs = 10,
  alignTo = "detected",
  invertAligned = FALSE
)
}
\arguments{
\item{dataset}{\link{msi.dataset-class} object. See \link{msiDataset}.}
}
\value{
\link{ms.image-class} object with binary intensities.
}
\description{
Calculate the binary reference image using k-means clustering. K-Means is run
on the first `npcs` principal components to speed up the calculations.
}
