% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_refimg.R
\name{refImageBinaryOtsu}
\alias{refImageBinaryOtsu}
\title{Calculate the binary reference image using Otsu's thresholding.}
\usage{
refImageBinaryOtsu(image)
}
\arguments{
\item{image}{\link{ms.image-class} object. See \link{msImage}.}
}
\value{
\link{ms.image-class} object with binary intensities.
#'
}
\description{
Calculate the binary reference image using Otsu's thresholding.
}
