% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msi.dataset_methods.R
\name{totalIonCountMSI,msi.dataset-method}
\alias{totalIonCountMSI,msi.dataset-method}
\alias{totalIonCountMSI}
\title{Generates an msImage representing pixels total-ion-counts. This
image can be used to qualitatively evaluate the spatial heterogeneity of the
sample.}
\usage{
\S4method{totalIonCountMSI}{msi.dataset}(object)
}
\arguments{
\item{object}{\link{msi.dataset-class} object.}
}
\value{
\link{ms.image-class} object representing the total ion counts.
}
\description{
Generates an msImage representing pixels total-ion-counts. This
image can be used to qualitatively evaluate the spatial heterogeneity of the
sample.
}
